Fix code style of Json::ArrAppend (#1846)

+++ src/types/redis_json.h
@@ -37,7 +37,7 @@ class Json : public Database {
   rocksdb::Status Get(const std::string &user_key, const std::vector<std::string> &paths, JsonValue *result);
   rocksdb::Status Type(const std::string &user_key, const std::string &path, std::vector<std::string> *results);
   rocksdb::Status ArrAppend(const std::string &user_key, const std::string &path,
-                            const std::vector<std::string> &values, std::vector<uint64_t> *result_count);
+                            const std::vector<std::string> &values, std::vector<size_t> *result_count);
 
  private:
   rocksdb::Status write(Slice ns_key, JsonMetadata *metadata, const JsonValue &json_val);
